requirements.txt
Flask==2.3.3
pyodbc==4.0.39
python-dotenv==1.0.0
google-generativeai==0.3.0
pdfplumber==0.10.3
Pillow==10.0.1
pytesseract==0.3.10
werkzeug==2.3.7